==> 03_sentiment_analysis.ipynb <==
# 03_Sentiment Analysis

### Import Libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from transformers import pipeline

### Define File Paths

fname_sentences  = 'data/sentences.csv'
fname_similarity = 'data/result_similarity.pkl'
fname_result     = 'data/result_both.pkl'

### Read Similarity Score File

df_s = pd.read_pickle(fname_similarity)
df_s

### Analyze Sentiment

classifier = pipeline('sentiment-analysis', model='distilbert-base-uncased-finetuned-sst-2-english')

sent_list = df_s['sentence'].tolist()

%%time

CNT = 100
LEN = len(sent_list)

result = []
for i in range(0, LEN, CNT):
    off_e = (i+CNT) if (i+CNT) < LEN else LEN
    
    print(f'sent_list[{i}:{off_e}]')
    # Max length of sentence: 512
    res = classifier(sent_list[i:off_e], truncation=True)
    
    result.extend(res)
        
print('==== End of jobs ====')

### Reverse Negative Scores

label = []
score = []
for dic in result:
    if dic['label'] == 'NEGATIVE':
        dic['score'] = 1 - dic['score']
    else:
        label.append(dic['label'])
    score.append(dic['score'])

df_s['label'] = label
df_s['score'] = score

df_s['label'] = df_s['label'].str.title()
df_s['score'] = df_s['score'].round(2)

df_s

### Pickle the Result 

df_s.to_pickle(fname_result)

---

# End of file
